backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
category_encoders==2.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
deskew==1.6.0
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.115.6
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
filelock==3.28.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
gdown==6.0.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
iopath==0.1.10
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
langdetect==1.0.9
layoutparser==0.3.4
lazy-loader==0.5
llvmlite==0.47.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.2
onnxruntime==1.24.4
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==3.0.1
parso==0.8.5
patsy==1.0.2
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.5.1
playwright==1.58.0
playwright-stealth==2.0.2
polars==1.39.3
polars-runtime-32==1.39.3
pooch==1.9.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pyee==13.0.1
Pygments==2.19.2
PyMatting==1.1.15
pyparsing==3.3.2
pypdfium2==5.7.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.17
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
rembg==2.0.75
requests==2.32.5
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.4.11
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.49
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.32.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
